refactor(navbar): split navbar into modular subcomponents and add profile avatar
- Add user avatar to desktop dropdown menu header
- Extract NavbarSearch subcomponent for desktop search bar
- Extract NavbarNotifications subcomponent to deduplicate notification logic (DRY)
- Extract NavbarUserMenu subcomponent for desktop user profile dropdown
- Extract NavbarMobileDrawer subcomponent for mobile navigation sheet
- Reduce navbar.tsx from ~1.1k lines to clean layout coordinator
